04/09/2025

+++ src/main/java/com/realestate/main/entity/Token.java
@@ -27,8 +27,7 @@ public class Token {
 	private LocalDate tokenDeadLine;
 	private String agentName;
 	private String finalStatus;
-	private double finalAmount;
-
+	private long finalAmount;
 	
 	@ManyToOne
 	@JsonBackReference(value = "lead-tokens")
